Add franc.eval.FilterInterface
This adds separates the minimum interface that is requried for
EvaluationRun from the minimum parameters required for all filtering
techniques.
